Python - Move files into a new package
