+++ lizmap/table_manager/__init__.py
@@ -1,0 +1,3 @@
+__copyright__ = 'Copyright 2023, 3Liz'
+__license__ = 'GPL version 3'
+__email__ = 'info@3liz.org'
