+++ lizmap/test/test_table_manager.py
@@ -18,8 +18,8 @@ from lizmap.definitions.locate_by_layer import LocateByLayerDefinitions
 from lizmap.definitions.time_manager import TimeManagerDefinitions
 from lizmap.definitions.tooltip import ToolTipDefinitions
 from lizmap.forms.atlas_edition import AtlasEditionDialog
-from lizmap.forms.table_manager import TableManager
 from lizmap.qgis_plugin_tools.tools.resources import plugin_test_data_path
+from lizmap.table_manager.table_manager import TableManager
 
 __copyright__ = 'Copyright 2020, 3Liz'
 __license__ = 'GPL version 3'
@@ -44,7 +44,7 @@ class TestTableManager(unittest.TestCase):
         definitions = FilterByFormDefinitions()
 
         table_manager = TableManager(
-            None, definitions, None, table, None, None, None, None, None)
+            None, definitions, None, table, None, None, None, None)
 
         json = {
             '0': {
@@ -128,7 +128,7 @@ class TestTableManager(unittest.TestCase):
     def test_form_filter_3_7(self):
         """ Test to write to 3.6 format. """
         table_manager = TableManager(
-            None, FilterByFormDefinitions(), None, QTableWidget(), None, None, None, None, None)
+            None, FilterByFormDefinitions(), None, QTableWidget(), None, None, None, None)
 
         json = {
             '0': {
@@ -188,7 +188,7 @@ class TestTableManager(unittest.TestCase):
         definitions = FilterByLoginDefinitions()
 
         table_manager = TableManager(
-            None, definitions, None, table, None, None, None, None, None)
+            None, definitions, None, table, None, None, None, None)
 
         json = {
             'lines': {
@@ -222,7 +222,7 @@ class TestTableManager(unittest.TestCase):
     def test_dataviz_definitions(self):
         """Test dataviz collections keys."""
         table_manager = TableManager(
-            None, DatavizDefinitions(), None, QTableWidget(), None, None, None, None, None)
+            None, DatavizDefinitions(), None, QTableWidget(), None, None, None, None)
         expected = [
             'type', 'title', 'description', 'layerId', 'x_field', 'aggregation',
             'traces', 'html_template', 'layout', 'popup_display_child_plot', 'stacked',
@@ -236,7 +236,7 @@ class TestTableManager(unittest.TestCase):
         definitions = DatavizDefinitions()
 
         table_manager = TableManager(
-            None, definitions, None, table, None, None, None, None, None)
+            None, definitions, None, table, None, None, None, None)
 
         # Lizmap 3.3
         json = {
@@ -289,7 +289,7 @@ class TestTableManager(unittest.TestCase):
         definitions = DatavizDefinitions()
 
         table_manager = TableManager(
-            None, definitions, None, table, None, None, None, None, None)
+            None, definitions, None, table, None, None, None, None)
 
         # Lizmap 3.3
         json = {
@@ -430,7 +430,7 @@ class TestTableManager(unittest.TestCase):
         definitions = DatavizDefinitions()
 
         table_manager = TableManager(
-            None, definitions, None, table, None, None, None, None, None)
+            None, definitions, None, table, None, None, None, None)
 
         # Lizmap 3.3
         json = {
@@ -579,7 +579,7 @@ class TestTableManager(unittest.TestCase):
         definitions = DatavizDefinitions()
 
         table_manager = TableManager(
-            None, definitions, None, table, None, None, None, None, None)
+            None, definitions, None, table, None, None, None, None)
 
         json = {
             '0': {
@@ -638,7 +638,7 @@ class TestTableManager(unittest.TestCase):
     def test_dataviz(self):
         """Test we can read dataviz 3.4 format."""
         table_manager = TableManager(
-            None, DatavizDefinitions(), None, QTableWidget(), None, None, None, None, None)
+            None, DatavizDefinitions(), None, QTableWidget(), None, None, None, None)
 
         json = {
             '0': {
@@ -722,7 +722,7 @@ class TestTableManager(unittest.TestCase):
     def test_filter_by_polygon(self):
         """ Test table manager with filter by polygon. """
         table_manager = TableManager(
-            None, FilterByPolygonDefinitions(), None, QTableWidget(), None, None, None, None, None)
+            None, FilterByPolygonDefinitions(), None, QTableWidget(), None, None, None, None)
 
         json = {
             'config': {
@@ -755,7 +755,7 @@ class TestTableManager(unittest.TestCase):
         definitions = ToolTipDefinitions()
 
         table_manager = TableManager(
-            None, definitions, None, table, None, None, None, None, None)
+            None, definitions, None, table, None, None, None, None)
 
         json = {
             'lines': {
@@ -779,7 +779,7 @@ class TestTableManager(unittest.TestCase):
         definitions = AttributeTableDefinitions()
 
         table_manager = TableManager(
-            None, definitions, None, table, None, None, None, None, None)
+            None, definitions, None, table, None, None, None, None)
 
         json = {
             'lines': {
@@ -808,7 +808,7 @@ class TestTableManager(unittest.TestCase):
         definitions = TimeManagerDefinitions()
 
         table_manager = TableManager(
-            None, definitions, None, table, None, None, None, None, None)
+            None, definitions, None, table, None, None, None, None)
 
         # JSON from 3.3 with all fields
         json = {
@@ -875,7 +875,7 @@ class TestTableManager(unittest.TestCase):
         definitions = EditionDefinitions()
 
         table_manager = TableManager(
-            None, definitions, None, table, None, None, None, None, None)
+            None, definitions, None, table, None, None, None, None)
 
         json = {
             'lines': {
@@ -958,7 +958,7 @@ class TestTableManager(unittest.TestCase):
         definitions = LocateByLayerDefinitions()
 
         table_manager = TableManager(
-            None, definitions, None, table, None, None, None, None, None)
+            None, definitions, None, table, None, None, None, None)
 
         json = {
             'lines': {
@@ -1039,7 +1039,7 @@ class TestTableManager(unittest.TestCase):
         definitions = AtlasDefinitions()
 
         table_manager = TableManager(
-            None, definitions, AtlasEditionDialog, table, None, None, None, None, None)
+            None, definitions, AtlasEditionDialog, table, None, None, None, None)
 
         self.assertEqual(table.columnCount(), len(definitions.layer_config.keys()))
 
@@ -1078,7 +1078,7 @@ class TestTableManager(unittest.TestCase):
         definitions._use_single_row = False
 
         table_manager = TableManager(
-            None, definitions, AtlasEditionDialog, table, None, None, None, None, None)
+            None, definitions, AtlasEditionDialog, table, None, None, None, None)
 
         self.assertEqual(table.columnCount(), len(definitions.layer_config.keys()))
 
@@ -1176,7 +1176,7 @@ class TestTableManager(unittest.TestCase):
         definitions = AtlasDefinitions()
         definitions._use_single_row = False
         table_manager = TableManager(
-            None, definitions, AtlasEditionDialog, table, None, None, None, None, None)
+            None, definitions, AtlasEditionDialog, table, None, None, None, None)
 
         json = {
             'layers': [
@@ -1233,7 +1233,7 @@ class TestTableManager(unittest.TestCase):
         }
 
         table_manager = TableManager(
-            None, definitions, AtlasEditionDialog, table, None, None, None, None, None)
+            None, definitions, AtlasEditionDialog, table, None, None, None, None)
 
         self.assertEqual(table_manager.table.rowCount(), 0)
         table_manager.from_json(json)
